+++ lib/rails_admin_import/eager_load.rb
@@ -1,0 +1,3 @@
+require "rails_admin_import"
+require "rchardet"
+require "simple_xlsx_reader"